Move code from prepare_proposal.rs to vote_extensions.rs
